+++ NEWS.md
@@ -1,7 +1,13 @@
 # shinycssloaders 0.1.1
 
-Fix Title-case for CRAN submission.
+## New features
 
+* Add proxy.height option to show a "placeholder" that contains the spinner for outputs of unknown height (e.g. tables with non-fixed heights).
+
+## Bug-fixes
+
+* Fix vertical scroll-bar appearing for Type 3 spinners
+ 
 # shinycssloaders 0.1.0
 
 The first working version of the package.